+++ 42cursus/ft_irc/includes/Client.hpp
@@ -9,7 +9,7 @@ class Client {
 		~Client(void);
 
 		// ./class/Client/Client.cpp
-		unsigned short getFd(void) const;
+		int getFd(void) const;
 		std::string getIp(void) const;
 		unsigned short getPort(void) const;
